requirements.txt
asgiref==3.4.1 
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
distlib==0.3.3
Django==4.0.1
django-auth-ldap==4.0.0
django-cors-headers==3.11.0
django-jalali-date==1.0.1
django-middleware-global-request==0.2.0
django-mssql==1.8
django-taggit==2.1.0
django-user-agents==0.4.0
django-utils-six==2.0
djangorestframework==3.13.1
filelock==3.4.0
idna==3.3
install==1.3.5
jdatetime==3.6.4
mssql-django==1.1
Pillow==9.0.0
platformdirs==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pyodbc==4.0.32
python-dotenv==0.20.0
./python_ldap-3.4.0-cp39-cp39-win_amd64.whl
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
virtualenv==20.10.0
